pyproject.toml
[tool.poetry]
name = "recognition-telegram-bot"
version = "0.1.0"
description = "Бот для Telegram"
authors = ["Korzunov Evgeniy <korzunov.e.v@gmail.com>"]
readme = "README.md"
packages = []

#pythonpath = [
#  "src"
#]

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.9.1"
isort = "^5.12.0"
autoflake = "^2.2.1"


[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.7"
aiohttp = "^3.8.6"
#ffmpeg-python = "^0.2.0"
#pydub = "^0.25.1"
pydantic-settings = "^2.0.3"
JSON-log-formatter = "^0.5.2"
python-logging-loki = "^0.3.1"
